+++ app/src/main/java/majere/raistlin/utils/AreaChecker.java
@@ -2,20 +2,20 @@ package majere.raistlin.utils;
 
 public class AreaChecker {
 
-    public static Boolean checkDot(Integer x, Float y, Integer r){
+    public static Boolean checkDot(Float x, Float y, Integer r){
         if (r == 0) {
             return false;
         }
 
 
         if (x >= 0 && y >= 0) {
-            return x <= r && y <= r / 2;
+            return x <= r && y <= (float) r / 2;
         }
         else if (x <= 0 && y <= 0) {
-            return (x * x + y * y) <= (r * r) / 4;
+            return (x * x + y * y) <= (float) (r * r) / 4;
         }
         else if (x <= 0 && y >= 0) {
-            return y <= (0.5*x + r/2);
+            return y <= (0.5*x + (float) r/2);
         }
         return false;
     }
